Log in defect fixed.
Fixed a defect where code did not recognize properly when not
logged in. Sector Alarm sometimes serves the default page, even
though not logged in, rather than redirect to login form.
Now we try to check status and if we get a 401, we assume that
we're not logged in and try to do so.